img_io4_incr.patch by Steve Ratcliffe
"I think we should remove the commented out code and I noticed
a few unneeded casts remaining (some maybe not actually as a result
of the patch), and some javadoc errors."


git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@4168 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/BufferedImgFileReader.java
@@ -104,7 +104,7 @@ public class BufferedImgFileReader implements ImgFileReader {
 	 * @return int sign-extended value that was read.
 	 */
 	public int get1s() throws ReadFailedException {
-		return (int)get();
+		return get();
 	}
 
 	/**
